requirements.txt
fastapi
uvicorn[standard]
sqlmodel
SQLAlchemy
jinja2
python-dotenv
psutil
